// src/app/models/sidebar.models.ts
export interface LiveOnX {
    id: string,
    title: string;
    avatar: string,
    host: string;
    audience: string;
  }
  
  export interface Trend {
    id: string,
    title: string;
    posts: string;
  }
  
  export interface WhoToFollow {
    id: string,
    name: string;
    handle: string;
    avatar: string;
  }
  